refactor : BuyerMainPage에서Footer 부분 컴포넌트로 분리
