+++ src/App.js
@@ -1,6 +1,6 @@
 import GlobalStyle from "./style/GlobalStyle";
 import AuthContextProvider from "./contexts/AuthContext";
-import BuyerMainPage from "./pages/BuyerMainPage";
+import BuyerMainPage from "./pages/BuyerMainPage/BuyerMainPage";
 import { BrowserRouter as Router } from "react-router-dom";
 
 function App() {
